move message events to its own routes folder
